+++ R/fit_gambin.R
@@ -24,20 +24,22 @@ core_message = function(cores) {
 #' distributions depend on sample size, abundances of different communities should be compared
 #' on equally large samples. The sample size can be set by the \code{subsample} parameter.
 #' To estimate \code{alpha} from a standardised sample, the function must be run several
-#' times; see the examples. The \code{no_of_components} parameter enables mutlimodal gambin
+#' times; see the examples. The \code{no_of_components} parameter enables multimodal gambin
 #' distributions to be fitted. For example, setting \code{no_of_components} equal to 2, the bimodal
 #' gambin model is fitted. When a multimodal gambin model is fitted (with g modes), the return values are the alpha
 #' parameters of the g different component distributions, the max octave values for the g component distributions 
 #' (as the max octave values for the g-1 component distributions are allowed to vary), and the and the weight parameter(s) 
 #' which denote the fraction of objects within each g component distribution. When fitting multimodal gambin models
-#' (particuarly on large datasets), the optimisation algorithm can be slow. In such cases, the process
+#' (particularly on large datasets), the optimisation algorithm can be slow. In such cases, the process
 #' can be speeded up by using the \code{cores} parameter to enable parallel computing.
 #' 
-#' The \code{plot} method creates a barplot showing the observed
-#' number of species in octaves, with the fitted GamBin distribution shown as black dots.
-#' @return The \code{fit_abunbances} function returns an object of class \code{gambin}, with the \code{alpha},
+#' The \code{plot} method creates a barplot showing the observed number of
+#' species in octaves, with the fitted GamBin distribution shown as black dots.
+#' The \code{summary.gambin} method provides additional useful information such
+#' as confidence intervals around the model parameter estimates.
+#' @return The \code{fit_abundances} function returns an object of class \code{gambin}, with the \code{alpha},
 #' \code{w}
-#' and \code{MaxOctave} parameters of the GamBin mixture distribution,
+#' and \code{MaxOctave} parameters of the gambin mixture distribution,
 #' the likelihood of the fit, and the empirical distribution over octaves.
 #' @importFrom stats optim
 #' @examples
@@ -71,16 +73,19 @@ fit_abundances <- function(abundances, subsample = 0, no_of_components = 1, core
   }
 
   if(no_of_components == 1) {
-    val = optim(par = 1, fn=ll_optim, maxoct = max(mydata$octave),
+    val = optim(par = 1, fn=ll_optim, maxoctave = max(mydata$octave),
                 values = mydata$octave, freq = mydata$species, method = "Brent",
                 lower = 0, upper = 100)
     val$octaves = max(mydata$octave)
+    ## Optim minimises so convert to negative
+    logLik = -val$value
   } else {
     core_message(cores)
     alpha = rep(1, no_of_components)
     w = rep(1/length(alpha), length(alpha) - 1)
     val = estimate_parameters(par = c(alpha, w), values = mydata$octave, 
                               freq = mydata$species, cores = cores)
+    logLik = -val$value
   }
   res = list()
   res$alpha = val$par[1:no_of_components]
@@ -89,7 +94,6 @@ fit_abundances <- function(abundances, subsample = 0, no_of_components = 1, core
   res$octaves = val$octaves
   res$convergence = val$convergence
   
-  logLik = -val$value
   ## -1: weights have to sum to 1
   ## -1: one of the octaves is equal to max(data)
   attr(logLik, "df") = no_of_components*2 + (no_of_components - 1) 
@@ -108,6 +112,16 @@ fit_abundances <- function(abundances, subsample = 0, no_of_components = 1, core
 
   res$max_octaves = res$max_octaves
   res$coefficients = c(alpha = res$alpha, w = res$w, max_octave = res$octaves)
+  
+  #get fitted values of each component distribution and the peak value of each for deconstruction function
+  if (no_of_components > 1) {
+    fvals <- mapply(function(x, y) dgambin_single(res$Data$octave, x, y) * 
+                                    sum(mydata$species), x = res$alpha, y = res$octaves)
+    res$peak_octave_individual <- apply(fvals, 2, which.max) - 1 #minus one as octaves start at 0
+  }
+  
+  #add information for use with summary function
+  res$cores <- cores
 
   attr(res, "nobs") = nrow(mydata)
   class(res) = "gambin"