+++ R/gambin-package.R
@@ -9,7 +9,7 @@
 #' and has been demonstrated to fit empirical data better than the most commonly used species-abundance models 
 #' (see references). Gambin is a stochastic model which combines the gamma distribution with a binomial sampling method. 
 #' To fit the gambin distribution, the abundance data is first binned into octaves. The expected abundance 
-#' octave of a species is given by the number of successfull consecutive Bernoulli 
+#' octave of a species is given by the number of successful consecutive Bernoulli 
 #' trials with a given parameter \code{p}. The parameter \code{p} of species is assumed to distributed 
 #' according to a gamma distribution. This approach can be viewed as linking the gamma distribution with 
 #' the probability of success in a binomial process with x trials. Use the \code{fit_abundances()} function to fit the 
@@ -19,7 +19,7 @@
 #' associated confidence intervals. Methods are provided for plotting the results, and for calculating the likelihood of fits.
 #' 
 #' The package now provides functionality to fit multimodal gambin distributions (i.e. a gambin distribution with 
-#' more than one mode).
+#' more than one mode), and to deconstruct and examine a multimodal gambin model fit (\code{deconstruct_modes}).
 #' 
 #' @references Matthews, T.J., Borregaard, M.K., Ugland, K.I., Borges, P.A.V, Rigal, F., Cardoso, P. and Whittaker, R.J. (2014) 
 #' The gambin model provides a superior fit to species abundance distributions with a single free parameter: evidence, implementation and interpretation. Ecography 37: 1002-1011.